issue counting
