+++ tools/issue-manager/package.json
@@ -12,6 +12,7 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "@octokit/types": "^9.0.0",
     "@pnp/core": "^3.12.1",
     "dotenv": "^16.0.3",
     "octokit": "^2.0.14"
