+++ tools/issue-manager/src/index.ts
@@ -1,5 +1,9 @@
+import { GetResponseDataTypeFromEndpointMethod } from "@octokit/types";
 import * as dotenv from "dotenv";
-import { dateAdd, combine } from "@pnp/core";
+// import { lockIssues } from "./lock-issues.js";
+
+import { writeFileSync } from "fs";
+
 import { Octokit } from "octokit";
 
 dotenv.config();
@@ -15,85 +19,78 @@ const { owner, repo, repo_url } = {
   repo_url: "https://github.com/OneDrive/onedrive-api-docs"
 };
 
+console.log(repo_url);
+
 // get all the issues in the repository, starting in pages of 100 moving from the oldest forwards
-const issues = await octokit.rest.issues.listForRepo({
-  owner,
-  repo,
-  state: "open",
-  filter: "repos",
-  per_page: 100,
-  direction: "asc"
-});
+// const issues = await octokit.rest.issues.listForRepo({
+//   owner,
+//   repo,
+//   state: "open",
+//   filter: "repos",
+//   per_page: 100,
+//   direction: "asc",
+// });
 
-const oneYearAgo = <Date>dateAdd(new Date(), "year", -1);
+// await lockIssues(octokit, owner, repo, repo_url, issues);
 
-console.log(`Cutoff Date: ${oneYearAgo.toLocaleDateString()}`);
+const issues: ListForRepoData = [];
 
-try {
+let page = 1;
 
-  for (let i = 0; i < issues.data.length; i++) {
+type ListForRepoData = GetResponseDataTypeFromEndpointMethod<typeof octokit.rest.issues.listForRepo>;
 
-    const issue = issues.data[i];
+while (true) {
 
-    const issueCreatedDate = new Date(issue.created_at);
-    const issueUrl = combine(repo_url, "/issues/", issue.number.toString());
+  const newIssues = await octokit.rest.issues.listForRepo({
+    owner,
+    repo,
+    state: "all",
+    filter: "repos",
+    per_page: 200,
+    page,
+  });
 
-    console.log(`Start: [${issueUrl}] ${issue.title}  ${issueCreatedDate.toLocaleDateString()}`);
-
-    if (issueCreatedDate < oneYearAgo) {
-
-      console.log("Beginning Processing");
-
-      // apply the automation label
-      console.log("Adding label");
-      await octokit.rest.issues.addLabels({
-        owner,
-        repo,
-        issue_number: issue.number,
-        labels: ["automation:Closed"]
-      });
-      console.log("Added label");
-
-      // update the issue with the automation comment
-      console.log("Creating comment");
-      await octokit.rest.issues.createComment({
-        owner,
-        repo,
-        issue_number: issue.number,
-        body: "As part of a repository clean up effort we are closing older issues. If this issue remains, please: open a new issue, reference this issue, and provide any additional details that may help in resolution. Thank you for your understanding as we work to improve our responsiveness.",
-      });
-      console.log("Created comment");
-
-      // update the issue as closed
-      console.log("Closing issue");
-      await octokit.rest.issues.update({
-        owner,
-        repo,
-        issue_number: issue.number,
-        state: "closed",
-      });
-      console.log("Closed issue");
-
-      // lock the issue
-      console.log("Locking issue");
-      await octokit.rest.issues.lock({
-        owner,
-        repo,
-        issue_number: issue.number,
-        lock_reason: "resolved"
-      });
-      console.log("Locked issue");
-
-    } else {
-
-      console.log("Issue is too new");
-    }
-
-    console.log(`End: [${issueUrl}]`);
+  if (newIssues === undefined || newIssues.data.length < 1) {
+    break;
   }
 
-} catch (e) {
+  const filteredIssues = newIssues.data.filter(i => i.pull_request === undefined);
 
-  // we just end on error
-  console.error(e);
+  issues.push(...filteredIssues);
+
+  if (page > 20) {
+    // here for safety
+    throw Error("Just looping.");
+  }
+
+  page++;
 }
+
+console.log(`Total Issues: ${issues.length}`);
+
+const issuesMatrix = new Map<number, number[]>();
+
+for (let i = 0; i < issues.length; i++) {
+
+  const issue = issues[i];
+
+  const d = new Date(issue.created_at);
+  const year = d.getFullYear();
+  const month = d.getMonth();
+
+  if (!issuesMatrix.has(year)) {
+    issuesMatrix.set(year, Array(12).fill(0));
+  }
+
+  issuesMatrix.get(year)![month]++;
+}
+
+const output: string[] = [];
+
+issuesMatrix.forEach((value, key) => {
+
+  output.push(`${key},${value.join(",")}`);
+
+});
+
+writeFileSync("test.csv", output.join("\n"));
